refactored by all remarks
