--- src/app.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { OrdersModule } from './modules/Orders/orders.module';
-
-@Module({
-  imports: [OrdersModule],
-  controllers: [AppController],
-  providers: [AppService],
-})
-export class AppModule {}
