+++ src/modules/app.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { AppController } from 'src/app.controller';
+import { AppService } from 'src/services/app.service';
+import { OrdersModule } from 'src/modules/Orders/orders.module';
+
+@Module({
+  imports: [OrdersModule],
+  controllers: [AppController],
+  providers: [AppService],
+})
+export class AppModule {}
